% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis2-package.R
\docType{package}
\name{usethis2-package}
\alias{usethis2}
\alias{usethis2-package}
\title{usethis2: Automate Analytic Project Setup and Development}
\description{
Automate analytic project setup tasks that are otherwise performed 
    manually. This includes setting up docker, spinning up a microservice, and 
    more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidylab.github.io/usethis2/}
  \item \url{https://github.com/tidylab/usethis2}
  \item Report bugs at \url{https://github.com/tidylab/usethis2/issues}
}

}
\author{
\strong{Maintainer}: Harel Lustiger \email{tidylab@gmail.com} (\href{https://orcid.org/0000-0003-2953-9598}{ORCID})

}
\keyword{internal}
